hw06/src/test/java/ru/otus/hw/repositories/JpaAuthorRepositoryImplTest.java
package ru.otus.hw.repositories;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.context.annotation.Import;
import ru.otus.hw.repositories.entities.Author;

import java.util.List;
import java.util.Optional;
import java.util.stream.IntStream;

import static org.assertj.core.api.Assertions.assertThat;

@DataJpaTest
@Import({JpaAuthorRepositoryImpl.class})
class JpaAuthorRepositoryImplTest {

    @Autowired
    private AuthorRepository repository;

    @MethodSource("getDbAuthors")
    @ParameterizedTest
    void findById_ShouldReturnCorrectAuthor_WhenHaveOriginal(Author expected) {
        Optional<Author> actual = repository.findById(expected.getId());

        assertThat(actual)
                .isPresent()
                .get()
                .usingRecursiveComparison()
                .isEqualTo(expected);
    }

    @Test
    void findAll_ShouldReturnCorrectAuthorsList_WhenHaveOriginalList() {
        List<Author> expected = getDbAuthors();
        List<Author> actual = repository.findAll();

        assertThat(actual)
                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("id")
                .containsExactlyInAnyOrderElementsOf(expected);
    }


    private static List<Author> getDbAuthors() {
        return IntStream.range(1, 4).boxed()
                .map(id -> new Author(id, "Author_" + id))
                .toList();
    }
}
